requirements.txt
dateparser
mistune @ git+https://github.com/alphatownsman/mistune.git@660b1c0100ecdd6cd6aca26a554be2606a67d67b
django~=3.2.16
django-markdownx @ git+https://github.com/alphatownsman/django-markdownx.git@e69480c64ad9c5d0499f4a8625da78cf2bb7691b
django-jsoneditor @ git+https://github.com/alphatownsman/django-jsoneditor.git@fa2ae41aeeb34447bd8a808a520e843c853fd16e
django-sass
rq>=1.12.0
django-rq
django-simple-history
django-hijack
django-user-messages
django-slack
django-ninja
django-polymorphic
django-maintenance-mode
django-tz-detect
django-bleach
django-redis
meilisearch
easy-thumbnails
lxml
openpyxl
psycopg2
requests
filetype
setproctitle
tqdm
opencc
dnspython
typesense
markdownify
igdb-api-v4
